LayoutTests/media/video-currentTime-set2-expected.txt

EVENT(canplaythrough)
EXPECTED (video.currentTime.toFixed(1) == '1.5') OK
EXPECTED (video.currentTime.toFixed(1) == '3.1') OK
END OF TEST
